use key pair to authenticate jwt

+++ src/Routes/V1/Boards/Router.ts
@@ -1,45 +1,29 @@
 import express, { Request, Response, NextFunction } from "express";
 import winston from "winston";
 import { v4 as uuidv4 } from "uuid";
-import jwt from "jsonwebtoken";
 import { body, validationResult, param, query } from "express-validator";
 import { Boards } from "./Boards";
+import { authenticate } from "Middlewares";
+import { AccessToken } from "Interface";
 
-const authenticate = (req: Request, res: Response, next: NextFunction) => {
-    const authHeader = req.headers.authorization;
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-        return res.status(401).json({ message: "Unauthorized" });
+function checkPermissions(
+    jwt: AccessToken,
+    action: "owns" | "edits" | "reads",
+    resource: "boards" | "catalogs" | "groups",
+    resourceId: string
+): boolean {
+    if (!jwt || !jwt[action] || !jwt[action]![resource]) {
+        return false;
     }
-    const token = authHeader.split(" ")[1];
+    return jwt[action]![resource].includes(resourceId);
+}
 
-    try {
-        const claims = jwt.verify(token, process.env.JWT_SECRET ?? "");
-        req.user = claims as { sub: string; boards?: Record<string, string[]> };
-        return next();
-    } catch (error) {
-        return res
-            .status(401)
-            .json({ message: "Invalid or expired JWT token" });
-    }
-};
-
-const checkBoardPermissions = (permission: string) => {
-    return (req: Request, res: Response, next: NextFunction) => {
-        const boardId = req.params.boardId || req.query.boardId;
-        if (
-            req.user &&
-            req.user.boards &&
-            req.user.boards[boardId]?.includes(permission)
-        ) {
-            next();
-        } else {
-            res.status(403).json({
-                message:
-                    "Forbidden - User does not have the necessary permissions for the resource",
-            });
-        }
-    };
-};
+function forbidden(res: Response): void {
+    res.status(403).json({
+        message:
+            "Forbidden - User does not have the necessary permissions for the resource",
+    });
+}
 
 export function getBoardsRouter(
     boards: Boards,
@@ -51,7 +35,7 @@ export function getBoardsRouter(
     router.post(
         "/boards",
         authenticate,
-        checkBoardPermissions("create"),
+        body("catalogId").isUUID(),
         body("title").optional().isString(),
         async (req: Request, res: Response) => {
             try {
@@ -62,6 +46,14 @@ export function getBoardsRouter(
 
                 const boardId = uuidv4();
                 const title: string = req.body.title || boardId;
+
+                const catalogId = req.params.catalogId;
+                if (
+                    !checkPermissions(req.user, "owns", "catalogs", catalogId)
+                ) {
+                    return forbidden(res);
+                }
+
                 await boards.createBoard(boardId, title);
                 return res.status(201).json({
                     boardId: boardId,
@@ -75,6 +67,7 @@ export function getBoardsRouter(
         }
     );
 
+    // Creating a new public board
     router.post(
         "/public-boards",
         body("title").optional().isString(),
@@ -103,13 +96,20 @@ export function getBoardsRouter(
     router.delete(
         "/boards",
         authenticate,
-        checkBoardPermissions("delete"),
+        body("catalogId").isUUID(),
         query("boardId").isUUID(),
         async (req: Request, res: Response) => {
             try {
                 const errors = validationResult(req);
                 if (!errors.isEmpty()) {
                     return res.status(400).json({ errors: errors.array() });
+                }
+
+                const catalogId = req.params.catalogId;
+                if (
+                    !checkPermissions(req.user, "owns", "catalogs", catalogId)
+                ) {
+                    return forbidden(res);
                 }
 
                 const boardId = req.query.boardId as string;
@@ -126,13 +126,20 @@ export function getBoardsRouter(
     router.post(
         "/boards/:boardId/duplicate",
         authenticate,
-        checkBoardPermissions("duplicate"),
+        body("catalogId").isUUID(),
         param("boardId").isUUID(),
         async (req: Request, res: Response) => {
             try {
                 const errors = validationResult(req);
                 if (!errors.isEmpty()) {
                     return res.status(400).json({ errors: errors.array() });
+                }
+
+                const catalogId = req.params.catalogId;
+                if (
+                    !checkPermissions(req.user, "owns", "catalogs", catalogId)
+                ) {
+                    return forbidden(res);
                 }
 
                 const originalBoardId = req.params.boardId;
@@ -150,7 +157,7 @@ export function getBoardsRouter(
     router.patch(
         "/boards/:boardId",
         authenticate,
-        checkBoardPermissions("edit"),
+        body("catalogId").isUUID(),
         param("boardId").isUUID(),
         body("newTitle").isString(),
         async (req: Request, res: Response) => {
@@ -161,7 +168,15 @@ export function getBoardsRouter(
                 }
 
                 const boardId = req.params.boardId;
+                const catalogId = req.params.catalogId;
                 const newTitle = req.body.newTitle;
+
+                if (
+                    !checkPermissions(req.user, "owns", "catalogs", catalogId)
+                ) {
+                    return forbidden(res);
+                }
+
                 await boards.renameBoard(boardId, newTitle);
                 return res.status(200).send();
             } catch (err) {
@@ -175,43 +190,11 @@ export function getBoardsRouter(
     router.post(
         "/boards/:boardId/events",
         authenticate,
-        checkBoardPermissions("edit"),
         param("boardId").isUUID(),
         body("eventId").isUUID(),
         body("eventBody").isObject(),
         async (req: Request, res: Response): Promise<any> => {
             try {
-                /*
-                const uuid = request.params.uuid;
-                const boardEvent = request.body;
-                const boardEventBody = boardEvent.body;
-                const board = await boards.getBoard(uuid);
-    	
-                if (!board) {
-                    logger.info(
-                        `post /api/v1/boards/:uuid/events Exception: find board ${uuid}`
-                    );
-                    response.status(404).end();
-                    return;
-                }
-                board
-                    .addEvent(boardEventBody.eventId, boardEventBody)
-                    .then((boardEvent) => {
-                        logger.info(
-                            `post /api/v1/boards/:uuid/events Success: add new event to ${uuid}, /n ${JSON.stringify(
-                                request.body,
-                                null,
-                                4
-                            )}`
-                        );
-    	
-                        ws.publish(
-                            board.uuid,
-                            new BoardEventSM(board.uuid, boardEvent)
-                        );
-                    });
-                response.end();
-                */
                 const errors = validationResult(req);
                 if (!errors.isEmpty()) {
                     return res.status(400).json({ errors: errors.array() });
@@ -220,6 +203,14 @@ export function getBoardsRouter(
                 const boardId = req.params.boardId;
                 const eventId = req.body.eventId;
                 const eventBody = req.body.eventBody;
+
+                if (
+                    !checkPermissions(req.user, "owns", "boards", boardId) &&
+                    !checkPermissions(req.user, "edits", "boards", boardId)
+                ) {
+                    return forbidden(res);
+                }
+
                 const boardEvent = await boards.addEventToBoard(
                     boardId,
                     eventId,
@@ -237,7 +228,6 @@ export function getBoardsRouter(
     router.get(
         "/boards/:boardId/events",
         authenticate,
-        checkBoardPermissions("view"),
         param("boardId").isUUID(),
         query("page").optional().isInt({ min: 1 }),
         query("limit").optional().isInt({ min: 1, max: 100 }),
@@ -251,6 +241,15 @@ export function getBoardsRouter(
                 const boardId = req.params.boardId;
                 const page: number = parseInt(req.query.page as string) || 1;
                 const limit: number = parseInt(req.query.limit as string) || 10;
+
+                if (
+                    !checkPermissions(req.user, "owns", "boards", boardId) &&
+                    !checkPermissions(req.user, "edits", "boards", boardId) &&
+                    !checkPermissions(req.user, "views", "boards", boardId)
+                ) {
+                    return forbidden(res);
+                }
+
                 const events = await boards.getBoardEvents(
                     boardId,
                     page,
@@ -268,7 +267,6 @@ export function getBoardsRouter(
     router.post(
         "/boards/:boardId/links",
         authenticate,
-        checkBoardPermissions("isOwner"),
         param("boardId").isUUID(),
         body("type").isIn(["read", "edit"]),
         async (req: Request, res: Response) => {
@@ -281,6 +279,11 @@ export function getBoardsRouter(
                 const boardId = req.params.boardId;
                 const type = req.body.type;
                 const linkId = uuidv4();
+
+                if (!checkPermissions(req.user, "owns", "boards", boardId)) {
+                    return forbidden(res);
+                }
+
                 await boards.createLink(boardId, type, linkId);
 
                 const linkUri = `./boards/${linkId}`;
@@ -299,7 +302,6 @@ export function getBoardsRouter(
     router.delete(
         "/boards/:boardId/links/:linkId",
         authenticate,
-        checkBoardPermissions("isOwner"),
         param("boardId").isUUID(),
         param("linkId").isUUID(),
         async (req: Request, res: Response) => {
@@ -311,6 +313,11 @@ export function getBoardsRouter(
 
                 const boardId = req.params.boardId;
                 const linkId = req.params.linkId;
+
+                if (!checkPermissions(req.user, "owns", "boards", boardId)) {
+                    return forbidden(res);
+                }
+
                 await boards.deleteLink(boardId, linkId);
 
                 return res.status(204).send();
